Refactor: Remove commented out code in main.cpp
Explanation: The changes in the provided git diff show that the commented out code related to generating XML output for the Jenkins CPPunit plugin has been removed from the main.cpp file. This can be considered as a refactoring change to clean up the code and remove unnecessary or unused code.