Fix printing staff in group function in SheetBuilder.cpp
In the SheetBuilder.cpp file, the printStaffInGroup() function has been modified to fix an issue with printing staff in a group.

Previously, the count variable was initialized to 1, but it should be initialized to the total number of staffs in the group. This has been fixed by calling the getStaffCount() function to retrieve the correct count value.

Additionally, a new char array, sheetStaff, has been declared to store the sheet name. The sheet name is then copied into sheetStaff using the strcpy() function.

Inside the for loop, a new string, newString, is created to hold the partial.ly file name for each staff. The string is constructed by concatenating "_Staff_", the count value converted to a string, and ".partial.ly". This new string is then appended to sheetStaff using the strcat() function.

Finally, the staff's sheet file is opened using the updated sheetStaff as the file name.

These changes ensure that the correct staff sheets are printed for each staff in the group.