Update sound design creation in mmw script
The `createSoundDesign` function in the `mmw` script has been modified to use the `modsyn` binary. The `modsyn` binary takes two arguments: the patch file and the name of the output file. The function now calls `modsyn` with the appropriate arguments and then uses `csound` to generate the final sound design file.

The changes made to the `mmw` script are as follows:

```diff
createSoundDesign(){
-  projOrc="
-  sr        = 48000
-  kr        = 4800
-  ksmps     = 10
-  nchnls    = 9
-  0dbfs     = 1
-
-  "
-  echo "$projOrc" > $PROJPATH/sound_design/$PROJ.orc
+
+  modsyn $1 $2
+
+  csound -d -o$PROJPATH/$PROJ.wav -F $PROJPATH/tmp/$PROJ.mid $PROJPATH/sound_design/$PROJ.csd
 }
```

Additionally, the `modsyn` binary has been updated to accept the name of the output file as a command-line argument. The changes made to the `modsyn` binary are as follows:

```diff
int main(int argc, char **argv){

   const char* patch = argv[1];
+  const char* name = argv[2];

-  ModSyn modsyn= ModSyn(patch);
+  ModSyn modsyn= ModSyn(patch,
+                        name);
   modsyn.processPatch();

   return 0;
}
```

The `ModSyn` class has also been updated to accept the name of the output file as a constructor argument. The changes made to the `ModSyn` class are as follows:

```diff
ModSyn::ModSyn(){};

-ModSyn::ModSyn(const char* patch){
+ModSyn::ModSyn(const char* patch,
+               const char* name){
+ this->patch= patch;
+ this->name= name;
};
```

These changes allow for the creation of sound design using the `modsyn` binary and the generation of the final sound design file using `csound`.