Refactor SheetBuilder.cpp to fix appending file issue
In the SheetBuilder.cpp file, there is an issue with appending files. The code is not correctly appending the staff group headers to the output file. To fix this issue, I made the following changes:

1. Removed the line `strcpy(sheetStaff, this->sheetName);` as it is not necessary.

2. Commented out the line `string newString = "_Staff_Group_Header_" + std::to_string(i) + ".ly";` and `strcat(sheetStaff, newString.c_str());` as they are not needed.

3. Modified the for loop to start from 1 and go up to the staff group count using `i <= this->getStaffGroupCount()`.

4. Added the line `strcpy(sheetStaff, this->sheetName);` inside the for loop to reset the `sheetStaff` variable for each iteration.

5. Added the line `appendFile(sheetStaff, outputFile);` inside the for loop to append the staff group header file to the output file.

After making these changes, the code should correctly append the staff group headers to the output file.