Refactor ModSyn class initialization and logging
In this commit, I refactored the initialization of the ModSyn class in the main.cpp file. I removed the commented out lines for closing files and freeing memory, as they were not necessary.

In the modsyn.cpp file, I made changes to the logging statements. I removed the initial log statement before opening the log file, as it was redundant. Instead, I added a new log statement to indicate the start of the Modular Synthesiser.

These changes improve the code readability and remove unnecessary code.